Test bare module worker
